refactor(module/control): 将BigDecimal改为Double类型
- 在ControlSettingGasDTO和ControlSettingWindDTO中，将BigDecimal字段改为Double类型
- 在相关服务实现类中，将BigDecimal转换为Double后再设置到DTO对象
- 这个改动可以简化数据结构，减少不必要的类型转换操作

+++ src/main/java/com/siact/module/control/dto/ControlSettingGasDTO.java
@@ -6,8 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -21,13 +19,13 @@ public class ControlSettingGasDTO {
     private String dataCode;
 
     @ApiModelProperty("dcs运行值")
-    private BigDecimal runningDcsVal;
+    private Double runningDcsVal;
 
     @ApiModelProperty("智控算法计算值")
-    private BigDecimal gasAlgorithmCalcVal;
+    private Double gasAlgorithmCalcVal;
 
     @ApiModelProperty("人工调整值")
-    private BigDecimal gasManualVal;
+    private Double gasManualVal;
 
     @ApiModelProperty("是否自动模式,1:是 0:否")
     private Boolean autoState;
